Refactor token handling in refresh token stores
Updated using directives in SqlServerRefreshTokenStore.cs to remove an unused import and add System.Data. Changed method signatures for Get in SqlServerRefreshTokenStore.cs and InMemoryRefreshTokenStore.cs to use 'token' instead of 'key'. Modified SQL query and in-memory search to use TokenString. Updated DefaultJwtAuthManager.cs to use a new GUID string as the key in AddOrUpdate.